AMD OpenMP capability

+++ vlp4d_mpi/openmp/Spline.hpp
@@ -394,10 +394,14 @@ namespace Spline {
     const int nvx_max = dom->local_nxmax_[2] + HALO_PTS + 1;
     const int nvy_max = dom->local_nxmax_[3] + HALO_PTS + 1;
 
-    const int nx  = nx_max - nx_min;
-    const int ny  = ny_max - ny_min;
-    const int nvx = nvx_max - nvx_min;
-    const int nvy = nvy_max - nvy_min;
+    const size_t nx  = static_cast<size_t>(nx_max - nx_min);
+    const size_t ny  = static_cast<size_t>(ny_max - ny_min);
+    const size_t nvx = static_cast<size_t>(nvx_max - nvx_min);
+    const size_t nvy = static_cast<size_t>(nvy_max - nvy_min);
+    //const int nx  = nx_max - nx_min;
+    //const int ny  = ny_max - ny_min;
+    //const int nvx = nvx_max - nvx_min;
+    //const int nvy = nvy_max - nvy_min;
 
     if(std::is_same_v<layout_type, stdex::layout_left>) {
       #if defined( ENABLE_OPENMP_OFFLOAD )
@@ -438,10 +442,10 @@ namespace Spline {
     const int ny_max = dom->local_nxmax_[1] + HALO_PTS + 1;
     const int nvx_max = dom->local_nxmax_[2] + HALO_PTS + 1;
     const int nvy_max = dom->local_nxmax_[3] + HALO_PTS + 1;
-    const int nx  = nx_max - nx_min;
-    const int ny  = ny_max - ny_min;
-    const int nvx = nvx_max - nvx_min;
-    const int nvy = nvy_max - nvy_min;
+    const size_t nx  = static_cast<size_t>(nx_max - nx_min);
+    const size_t ny  = static_cast<size_t>(ny_max - ny_min);
+    const size_t nvx = static_cast<size_t>(nvx_max - nvx_min);
+    const size_t nvy = static_cast<size_t>(nvy_max - nvy_min);
 
     if(std::is_same_v<layout_type, stdex::layout_left>) {
       #if defined( ENABLE_OPENMP_OFFLOAD )
